# Remove pure function annotation