--- a/src/dual.jl
+++ b/src/dual.jl
@@ -337,7 +337,7 @@ end
 # Promotion/Conversion #
 ########################
 
-Base.@pure function Base.promote_rule(::Type{Dual{T1,V1,N1}},
+function Base.promote_rule(::Type{Dual{T1,V1,N1}},
                                       ::Type{Dual{T2,V2,N2}}) where {T1,V1,N1,T2,V2,N2}
     # V1 and V2 might themselves be Dual types
     if T2 ≺ T1