@KristofferC 2021-03-05
> This assumption seems to be violated here. 

Which part? There doesn't seem to be any generic functions being called, or?

Do you have an example of the stackoverflow problem?