@mohamed82008 2021-03-05
>Do you have an example of the stackoverflow problem?

It's my fault. But the error looks interesting when using `@pure`.
```julia
using ForwardDiff: ForwardDiff, Tag

struct MyTag end
ForwardDiff.:≺(::Type{Tag{F1,V1}}, ::Type{Tag{MyTag,V2}}) where {F1, V1, V2} = true
# Uncomment the next line to fix the error
#ForwardDiff.:≺(::Type{Tag{MyTag,V2}}, ::Type{Tag{F1,V1}}) where {F1, V1, V2} = false

dual1 = ForwardDiff.GradientConfig(MyTag(), [1.0]).duals[1]
dual2 = ForwardDiff.GradientConfig(sum, [1.0]).duals[1]

promote_type(typeof(dual1), typeof(dual2))
```