@andreasnoack 2021-03-05
I guess the assumption here is that `≺` is not extended.