@KristofferC 2021-03-06
So is there really a problem here? If you overload an internal function like this, well, caveat emptor.