@mohamed82008 2021-03-07
There is a problem that sometimes you do need to overload this internal function to specify order between custom tags. So the purity assumption is violated in a valid use case.