@KristofferC 2021-03-07
Presumably the pure was there for a reason. Did you check that no longer applies?