@mohamed82008 2021-03-07
The pure seems to have been introduced in https://github.com/JuliaDiff/ForwardDiff.jl/pull/247 back in 2017 as part of a big change to tackle perturbation confusion. It's hard to guess the intentions of @simonbyrne or @jrevels back then especially given that it was a very different Julia version and compiler back then.