@KristofferC 2021-03-08
So you are saying that the compiler has improved enough that the `@pure` is not needed? That might very well be possible, but at least some effort to confirm this would be beneficial, in my opinion.