@mohamed82008 2021-03-08
>So you are saying that the compiler has improved enough that the `@pure` is not needed? That might very well be possible, but at least some effort to confirm this would be beneficial, in my opinion.

I don't know if the compiler has improved enough that pure is not needed because I don't know why it was needed to begin with. There was no justification provided back then. Julia 0.6 was used back then. If you have a use case that could have benefited from `@pure` in Julia 0.6, I would be happy to check it in Julia 1.5/1.6 to see if `@pure` makes any difference at all.