@jrevels 2021-03-08
it's been a very long time since I looked at a ForwardDiff PR 😁 

> If you have a use case that could have benefited from `@pure` in Julia 0.6, I would be happy to check it in Julia 1.5/1.6 to see if `@pure` makes any difference at all.

I wouldn't be surprised if that were added back then to avoid potential perf issues w.r.t. calling https://github.com/JuliaDiff/ForwardDiff.jl/pull/247/files#diff-aeb036d3acc3ca0b52368fdcc3dbc8bca465b98bc269a1a5b4668b1dcb2e2f7eR12 and https://github.com/JuliaDiff/ForwardDiff.jl/pull/247/files#diff-aeb036d3acc3ca0b52368fdcc3dbc8bca465b98bc269a1a5b4668b1dcb2e2f7eR21 

Given the compiler advances since then, it would be great (and unsurprising) if it's no longer useful, but as @KristofferC mentioned, this is probably worth benchmarking in a nested differentiation case that hits this codepath to confirm it won't induce regressions in common cases