@simonbyrne 2021-03-08
I too suspect it would now work without it.