@KristofferC 2021-03-08
> If you have a use case that could have benefited from @pure in Julia 0.6, I would be happy to check it in Julia 1.5/1.6 to see if @pure makes any difference at all.

That's not really how things work. When you go in and remove a performance annotation it is on you to put at least some level of effort to show that the annotation is not needed anymore. For example, a benchmark where the function is used or looking at the generated code with the annotation there and removed.