@mohamed82008 2021-03-08
I used the following code to benchmark:
```julia
using ForwardDiff, BenchmarkTools

f(x, y) = ForwardDiff.derivative(y -> ForwardDiff.derivative(x -> (x + y)^2, x), y)
g(x, y) = ForwardDiff.jacobian(y -> ForwardDiff.gradient(x -> sum(x + y)^2, x), y)
h(x, y, z) = ForwardDiff.derivative(z -> ForwardDiff.derivative(y -> ForwardDiff.derivative(x -> (x + y + z)^2, x), y), z)

@btime f(2.0, 2.0);
@btime g($(ones(10)), $(ones(10)));
@btime h(1.0, 2.0, 3.0);
```

- With `@pure`:
```julia
julia> @btime f(2.0, 2.0);
  0.001 ns (0 allocations: 0 bytes)

julia> @btime g($(ones(10)), $(ones(10)));
  3.896 μs (15 allocations: 16.83 KiB)

julia> @btime h(1.0, 2.0, 3.0);
  1.500 ns (0 allocations: 0 bytes)
```

- Without `@pure`:
```julia
julia> @btime f(2.0, 2.0);
  0.001 ns (0 allocations: 0 bytes)

julia> @btime g($(ones(10)), $(ones(10)));
  3.854 μs (15 allocations: 16.83 KiB)

julia> @btime h(1.0, 2.0, 3.0);
  1.500 ns (0 allocations: 0 bytes)
```

The second number changes a fair bit every time I run it with and without pure but there seems to be no difference.

The code native generated in both cases:
- With `@pure`
```julia
julia> @code_native h(1.0, 2.0, 3.0)
	.section	__TEXT,__text,regular,pure_instructions
; ┌ @ REPL[61]:1 within `h'
	subq	$120, %rsp
; │┌ @ derivative.jl:14 within `derivative'
; ││┌ @ REPL[61]:1 within `#137'
; │││┌ @ derivative.jl:14 within `derivative'
; ││││┌ @ REPL[61]:1 within `#138'
	movsd	%xmm1, 24(%rsp)
	movabsq	$4607182418800017408, %rax      ## imm = 0x3FF0000000000000
	movq	%rax, 32(%rsp)
	movsd	%xmm2, 40(%rsp)
	movq	%rax, 48(%rsp)
; │││││┌ @ derivative.jl:14 within `derivative'
; ││││││┌ @ dual.jl:66 within `Dual' @ dual.jl:62 @ dual.jl:55 @ dual.jl:19
	movsd	%xmm0, 8(%rsp)
	movq	%rax, 16(%rsp)
; ││││││└
	movabsq	$"#139", %rax
	leaq	56(%rsp), %rdi
	leaq	24(%rsp), %rsi
	leaq	8(%rsp), %rdx
	callq	*%rax
; │└└└└└
	movsd	112(%rsp), %xmm0                ## xmm0 = mem[0],zero
	addq	$120, %rsp
	retq
	nopw	%cs:(%rax,%rax)
	nop
; └
```

- Without `@pure`
```julia
julia> @code_native h(1.0, 2.0, 3.0)
	.section	__TEXT,__text,regular,pure_instructions
; ┌ @ REPL[61]:1 within `h'
	subq	$120, %rsp
; │┌ @ derivative.jl:14 within `derivative'
; ││┌ @ REPL[61]:1 within `#137'
; │││┌ @ derivative.jl:14 within `derivative'
; ││││┌ @ REPL[61]:1 within `#138'
	movsd	%xmm1, 24(%rsp)
	movabsq	$4607182418800017408, %rax      ## imm = 0x3FF0000000000000
	movq	%rax, 32(%rsp)
	movsd	%xmm2, 40(%rsp)
	movq	%rax, 48(%rsp)
; │││││┌ @ derivative.jl:14 within `derivative'
; ││││││┌ @ dual.jl:66 within `Dual' @ dual.jl:62 @ dual.jl:55 @ dual.jl:19
	movsd	%xmm0, 8(%rsp)
	movq	%rax, 16(%rsp)
; ││││││└
	movabsq	$"#139", %rax
	leaq	56(%rsp), %rdi
	leaq	24(%rsp), %rsi
	leaq	8(%rsp), %rdx
	callq	*%rax
; │└└└└└
	movsd	112(%rsp), %xmm0                ## xmm0 = mem[0],zero
	addq	$120, %rsp
	retq
	nopw	%cs:(%rax,%rax)
	nop
; └
```

Looks identical to me.